Class dependent refactor
